Split appsec telemetry file into rasp, waf and user
